- Added warning to nuke.sh before running to ensure generated files are not deleted accidentally

+++ nuke.sh
@@ -1,7 +1,19 @@
 #!/bin/sh
 
 # Nukes all containers and deletes all generated files
-sh owner_key_src/nuke.sh
-sh pool_cert_src/nuke.sh
-rm -rf owner_key
-rm -rf pool_cert
+echo ""
+echo " ------------------------------- WARNING ------------------------------- "
+echo " - This operation will delete all local generated files and folders    - "
+echo " - Backup any files, folders or generated keys before running this     - "
+echo " - operation                                                           - "
+echo " ----------------------------------------------------------------------- "
+echo ""
+echo "Continue Nuke (y/n)? "
+read answer
+
+if [ "$answer" != "${answer#[Yy]}" ] ;then
+    sh owner_key_src/nuke.sh
+    sh pool_cert_src/nuke.sh
+    rm -rf owner_key
+    rm -rf pool_cert
+fi
